pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-atomic-operation"
version = "0.2.0"
description = "A YAML-driven atomic operation framework for robotic manipulation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenGHz" },
]
dependencies = [
  "hydra-core",
  "imageio",
  "msgpack",
  "msgpack-numpy",
  "numpy",
  "pydantic>=2",
  "pillow",
  "rpyc",
  "websockets",
]
urls = { "repository" = "https://github.com/OpenGHz/auto-atomic-operation.git" }

[project.scripts]
aao_demo = "auto_atom.runner.demo:main"
aao_eval = "auto_atom.runner.policy_eval:main"

[project.optional-dependencies]
mujoco = ["mujoco"]
gs = ["gaussian_renderer @ git+https://github.com/OpenGHz/GaussianRenderer.git", "gsplat", "torch", "scipy", "plyfile", "ninja"]


[tool.setuptools.packages.find]
include = ["auto_atom*"]
namespaces = true
